Modify the self-referencing Parsed struct to use the ouroboros crate.
Removing this usage of unsafe rust is aimed at lowering the maintenance
burden and risk of potential future memory-safety bugs. Future changes
to this part of the code won't need to be (re)analyzed manually to
confirm that the safety guarantees continue to hold, and it will be
easier for third-party organisations to be confident that askama is
memory-safe.